Move LayersScraper
