+++ uoftscrapers/scrapers/scraper/__init__.py
@@ -2,6 +2,7 @@ import os
 import shutil
 import logging
 
+
 class Scraper:
     """Scraper superclass."""
